# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/user": {
            "get": {
                "description": "This API Finds All Users",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "This API Finds All Users",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.FPUser"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.FPAPIResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.FPAPIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.FPAPIResponse"
                        }
                    }
                }
            }
        },
        "/user/add": {
            "post": {
                "description": "This API Creates a new User",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "This API Creates a new User",
                "parameters": [
                    {
                        "description": "New User",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.FPUser"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.FPUser"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.FPAPIResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.FPAPIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.FPAPIResponse"
                        }
                    }
                }
            }
        },
        "/user/delete/{userId}": {
            "delete": {
                "description": "This API deletes a User",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "This API deletes a User",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The id of a user",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.FPAPIResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.FPAPIResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.FPAPIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.FPAPIResponse"
                        }
                    }
                }
            }
        },
        "/user/find/{userId}": {
            "get": {
                "description": "This API Finds User by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "This API Finds User by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The id of a user",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.FPUser"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.FPAPIResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.FPAPIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.FPAPIResponse"
                        }
                    }
                }
            }
        },
        "/user/update": {
            "post": {
                "description": "This API Updates a User",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "This API Updates a User",
                "parameters": [
                    {
                        "description": "Existing User",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.FPUser"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.FPUser"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.FPAPIResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.FPAPIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.FPAPIResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "dto.FPAPIResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "statusCode": {
                    "type": "integer"
                }
            }
        },
...
        "dto.FPUser": {
            "type": "object",
            "properties": {
                "urs_firstname": {
                    "type": "string"
                },
                "usr_address": {
                    "type": "string"
                },
                "usr_email": {
                    "type": "string"
                },
                "usr_id": {
                    "type": "string"
                },
                "usr_idnumber": {
                    "type": "string"
                },
                "usr_middle_name": {
                    "type": "string"
                },
                "usr_password": {
                    "type": "string"
                },
                "usr_phonenumber": {
                    "type": "string"
                },
                "usr_role": {
                    "type": "string"
                },
                "usr_sp_id": {
                    "type": "string"
                },
                "usr_status": {
                    "type": "string"
                },
                "usr_surname": {
                    "type": "string"
                },
                "usr_username": {
                    "type": "string"
                }
            }
        },
...
        "models.FPUser": {
            "type": "object",
            "properties": {
                "fpuserAddress": {
                    "type": "string"
                },
                "fpuserEmail": {
                    "type": "string"
                },
                "fpuserFirstName": {
                    "type": "string"
                },
                "fpuserID": {
                    "type": "string"
                },
                "fpuserIDNumber": {
                    "type": "string"
                },
                "fpuserMiddleName": {
                    "type": "string"
                },
                "fpuserPassword": {
                    "type": "string"
                },
                "fpuserPhoneNumber": {
                    "type": "string"
                },
                "fpuserRole": {
                    "type": "string"
                },
                "fpuserSpID": {
                    "type": "string"
                },
                "fpuserStatus": {
                    "type": "string"
                },
                "fpuserSurname": {
                    "type": "string"
                },
                "fpuserusername": {
                    "type": "string"
                }
            }
        }
    }
}